+++ CommonConcepts/CommonConcepts.Test/ComputedFromTest.cs
@@ -336,7 +336,7 @@ namespace CommonConcepts.Test
             }
         }
 
-        private static readonly Regex FindRecomputesRegex = new Regex(@"GenericRepository\((?<module>.+)\.(?<entity>.+)\).*InsertOrUpdateOrDelete: Diff \((?<new>.+) new items, (?<old>.+) old items, (?<ins>.+) to insert, (?<upd>.+) to update, (?<del>.+) to delete\)");
+        private static readonly Regex FindRecomputesRegex = new Regex(@"Performance\.(?<module>\w+)\.(?<entity>\w+).*DiffFrom\w+: (?<new>.+) new items, (?<old>.+) old items, (?<ins>.+) to insert, (?<upd>.+) to update, (?<del>.+) to delete");
 
         private List<string> FindRecomputes(List<string> log)
         {